Resolve the Qodana findings for 2.0
Work through the remaining JetBrains inspections across src and tests:
inverted ifs, pattern merges, collection expressions, the field keyword,
System.Threading.Lock gates, async overloads where ordering is unchanged,
LibraryImport where marshalling is unchanged, redundant casts and defaults,
and dead private and internal code confirmed unused across the repository.

Test namespaces now follow their folders. The settings-page row styles move
to Themes/SettingsPages.axaml so the pages and the analyzer both see them,
and the two DataContext bindings that shared an element with x:DataType
move to a wrapping panel. The desktop action queue no longer posts an
async void lambda.

Two findings change behaviour: glyph SVG fill rules are now applied, and a
failed installer-exit wait logs a warning instead of shutting WSGM down.

Findings left open are deliberate: null checks on JSON, plugin and
interop data, public SDK contract members, COM vtable slots, XAML event
handlers, comparer collections that need C# 15, and intentional closures.
Qodana no longer checks the Windows manifests against an XML schema or
runs its Process.Start taint rule. The Python and JavaScript scans are
fixed as well.

+++ src/WSGM/Core/RtssFrametimeReader.cs
@@ -186,7 +186,7 @@ internal sealed class RtssFrametimeReader : IFrametimeSource, IDisposable
         var name = new byte[EntryNameLength];
         for (var index = 0; index < count; index++)
         {
-            var entry = arrayOffset + (long)index * entrySize;
+            var entry = arrayOffset + index * entrySize;
             if (entry < 0 || entry + entrySize > capacity)
             {
                 break;
